Update config schemas (#137)
* Update config schemas

This patch introduces a small refactor to update the configuration schemas in separate files. It also add the sqlalchemy docs for intersphinx.

* Update tests for schemas
