+++ command_line_assistant/config/schemas/__init__.py
@@ -1,0 +1,1 @@
+"""Module to hold the config schema and it's sub schemas."""
