+++ docs/source/config/schemas/backend.rst
@@ -1,0 +1,8 @@
+Backend
+=======
+
+.. automodule:: command_line_assistant.config.schemas.backend
+   :members:
+   :undoc-members:
+   :private-members:
+   :no-index:
