+++ docs/source/config/schemas/history.rst
@@ -1,7 +1,7 @@
-Schemas
+History
 =======
 
-.. automodule:: command_line_assistant.config.schemas
+.. automodule:: command_line_assistant.config.schemas.history
    :members:
    :undoc-members:
    :private-members:
