+++ docs/source/config/schemas/index.rst
@@ -1,0 +1,10 @@
+Schemas
+=======
+
+.. toctree::
+   :maxdepth: 2
+
+   backend
+   history
+   logging
+   output
