+++ docs/source/config/schemas/logging.rst
@@ -1,0 +1,8 @@
+Logging
+=======
+
+.. automodule:: command_line_assistant.config.schemas.logging
+   :members:
+   :undoc-members:
+   :private-members:
+   :no-index:
