+++ docs/source/config/schemas/output.rst
@@ -1,0 +1,8 @@
+Output
+======
+
+.. automodule:: command_line_assistant.config.schemas.output
+   :members:
+   :undoc-members:
+   :private-members:
+   :no-index:
