+++ tests/config/schemas/test_schema.py
@@ -1,0 +1,22 @@
+import pytest
+
+from command_line_assistant.config.schemas.backend import AuthSchema, BackendSchema
+from command_line_assistant.config.schemas.history import DatabaseSchema, HistorySchema
+from command_line_assistant.config.schemas.logging import LoggingSchema
+from command_line_assistant.config.schemas.output import OutputSchema
+
+
+@pytest.mark.parametrize(
+    ("schema",),
+    (
+        (LoggingSchema,),
+        (OutputSchema,),
+        (BackendSchema,),
+        (HistorySchema,),
+        (AuthSchema,),
+        (DatabaseSchema,),
+    ),
+)
+def test_initialize_schemas_default_values(schema):
+    # Making sure that we don't error out while initializing the schema with default values.
+    assert isinstance(schema(), schema)
